Google_ADK_A2A/a2a-project-main/mark_agent/pyproject.toml
[project]
name = "mark-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.25",
    "crewai[google-genai]>=1.9.3",
    "lancedb==0.3.0",
    "python-dotenv>=1.1.1",
]
